ai-tutor-backend/requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
python-multipart==0.0.20
google-genai==1.47.0
supabase==2.27.0
PyMuPDF==1.26.5
Pillow==11.3.0
python-dotenv